requirements.txt
alembic==1.4.2
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
Authlib==0.14.3
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.25
beautifulsoup4==4.8.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
ecdsa==0.14.1
elementpath==1.4.4
env==0.1.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.9
imutils==0.5.3
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
msrest==0.6.18
mysql-connector-python==8.0.20
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.4.0.44
pandas==1.0.3
Pillow==8.0.1
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pychalk==2.0.1
pycombat==0.14
pycparser==2.20
pylint==2.4.4
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-http-client==3.2.7
python-jose==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
sendgrid==6.4.5
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
tqdm==4.46.0
urllib3==1.25.8
virtualenv==20.0.20
waitress==1.4.3
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.11.2
